Frontend Component Cleanup - Admin page
The Admin buttons have been merged with the "Get me the next sentence" button. This is clearly too much of a scope for a simple button, especially as Admin features should reside on their own page anyway.

Changed include:
- Separate out the Admin features to a separate page
- Clean up the user session to make sure it stores whether you are an admin or not, so there's no need for separate check calls
- Fix up the cookie session security - currently we have been signing them with the default key that is inseucre so anyone could have modified the cookies to make themselves an Admin
