+++ .env.example
@@ -6,3 +6,4 @@ DB_USER=postgres
 DB_PASSWORD=0000
 DB_NAME=postgres
 DB_PORT=5432
+COOKIE_SECRET=e39a721e-f9fb-4a0c-beff-eb1f02b9c7d0
