+++ client/src/App.js
@@ -1,5 +1,6 @@
 import { Route, Routes } from "react-router-dom";
 import { useEffect, useState } from "react";
+import Admin from "./pages/Admin";
 import About from "./pages/About";
 import Home from "./pages/Home";
 import Login from "./pages/Login";
@@ -39,6 +40,7 @@ const App = () => {
 			<Route path="/" element={<Home user={user} />} />
 			<Route path="/about/this/site" element={<About />} />
 			<Route path="/login" element={<Login />} />
+			<Route path="/admin" element={<Admin user={user} />} />
 		</Routes>
 	);
 };
